requirements.txt
torch
torchvision
transformers
scikit-learn
matplotlib
seaborn
pandas
numpy
pillow
python-pptx
gradio
kaggle
